backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
python-jose[cryptography]
passlib
bcrypt
pydantic
python-dotenv
python-multipart

